requirements.txt
# OpenClaw Display Command Center Dependencies

# Display and image rendering
Pillow>=9.0.0

# WebSocket client for OpenClaw connection
websockets>=10.0

# .env file support for configuration
python-dotenv>=1.0.0

# Cryptography for OpenClaw device authentication
cryptography>=41.0.0

# Hardware dependencies (Raspberry Pi only)
spidev>=3.0
RPi.GPIO>=0.7.0

# I2C LCD support (16x2 character display)
RPLCD>=1.3.1
smbus2>=0.4.0
